backend/services/document-processor/bin/worker/requirements.txt
pulsar-client==3.4.0
asyncpg==0.29.0
aioboto3==12.4.0
dotenv